refactor: rename routes to follow lowercase convention
