+++ frontend/src/lib/components/ModuleRow.svelte
@@ -1,8 +1,15 @@
 <script lang="ts">
+    import {navigate} from "svelte-routing"
+
     export let moduleName: string;
     export let dueDate: string;
     export let progress: number;
     export let description: string;
+
+    function runModule(){
+        //router.redirect("/LearningOutcomePage")
+        navigate("/learning-outcomes", {replace: true})
+    }
 </script>
 
 <div class="grid-flow-col gap-8 my-4">
@@ -24,7 +31,7 @@
             <h2><b> {description}
             </b></h2>
             <div class = "text-right">
-                <button class="btn-highlight btn-icon">
+                <button on:click={runModule} class="btn-highlight btn-icon">
                     <div class="svgicon">
                         <svg xmlns="http://www.w3.org/2000/svg" width="80" height="60" fill="currentColor" class="bi bi-arrow-right-circle" viewBox="0 0 16 16">
                             <path fill-rule="evenodd" d="M1 8a7 7 0 1 0 14 0A7 7 0 0 0 1 8zm15 0A8 8 0 1 1 0 8a8 8 0 0 1 16 0zM4.5 7.5a.5.5 0 0 0 0 1h5.793l-2.147 2.146a.5.5 0 0 0 .708.708l3-3a.5.5 0 0 0 0-.708l-3-3a.5.5 0 1 0-.708.708L10.293 7.5H4.5z"/>
