+++ frontend/src/routes/chatbot-simulation.svelte
