+++ frontend/src/routes/index.svelte
@@ -1,9 +1,9 @@
 <script lang='ts'>
   import {Router, Route} from "svelte-routing"
 
-  import ModulePage from './ModulePage.svelte';
-  import LearningOutcomePage from './LearningOutcomePage.svelte';
-  import MessagingPage from './MessagingPage.svelte';
+  import ModulePage from './modules.svelte';
+  import LearningOutcomePage from './learning-outcomes.svelte';
+  import MessagingPage from './chatbot-simulation.svelte';
 </script>
